+++ packages/kevin_flutter_core/kevin_flutter_core_ios/lib/src/entity/theme/kevin_ui_font_entity_json.dart
@@ -1,6 +1,0 @@
-part of 'kevin_ui_font_entity.dart';
-
-Map<String, dynamic> _toJson(KevinUiFontEntity instance) => <String, dynamic>{
-      'size': instance.size,
-      'weight': instance.weight,
-    };