+++ packages/kevin_flutter_core/kevin_flutter_core_ios/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.3
+
+* **FIX**: ios theme color parsing
+* Upgrade kevin_ios to 2.2.3
+* Upgrade kevin_flutter_core_platform_interface to 1.0.3
+* Remove entity classes
+
 ## 1.0.2
 
 * Adjust repository/homepage URLs in pubspec.yaml